Extract blank value checker from attribute
